+++ recfm-java-addon/src/main/java/io/github/epi155/recfm/type/FieldEmbedGroup.java
@@ -29,6 +29,7 @@ public class FieldEmbedGroup extends NakedField implements EmbModel {
                 if (fld instanceof ParentFields) {
                     ((ParentFields) fld).getFields().forEach(this::shift);
                 }
+                ((NakedField) fld).setEmbedded(true);
                 return fld;
             }
         }